refactor: Remove ExpandedRailDestination and update navigation rail structure
