+++ lib/src/navigation_rail/navigation_rail.dart
@@ -11,8 +11,8 @@ import "../destination/destination_build_data.dart";
 import "../destination/destination_surface_strategy.dart";
 import "../material.dart";
 import "../navigation_icon.dart";
-import "expanded/navigation_rail_animation.dart";
 import "navigation_rail_theme.dart";
+import "widgets/navigation_rail_animation.dart";
 
 part "expanded/navigation_rail_destination.dart";
 part "widgets/indicator_ink_well.dart";
