+++ lib/src/navigation_rail/widgets/navigation_rail_animation.dart
